+++ packages/reference/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.16.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-reference@5.16.1...@contentful/field-editor-reference@5.16.2) (2023-09-28)
+
+**Note:** Version bump only for package @contentful/field-editor-reference
+
 ## [5.16.1](https://github.com/contentful/field-editors/compare/@contentful/field-editor-reference@5.16.0...@contentful/field-editor-reference@5.16.1) (2023-09-14)
 
 ### Bug Fixes